Created an image file, and added the four faction leaders as well as the logo for Rebel Monkz.
Styled the main background of the document with a
radial-gradient, while implementing a gray background for the faction
leaders section of the web page.

Created adequate spacing throughout the document with padding and margin

Made multiple changes throughout this journey without committing &
pushing to Github as much as one should. This was my first project and
got into a groove most of the time.